requirements.txt
apache-airflow-providers-microsoft-azure==5.0.1